Replace MyFCDashScope with OpenAILike class
