+++ src/pai_rag/modules/tool/utils.py
@@ -109,28 +109,31 @@ def get_weather_tools(config):
 def get_customized_tools(config):
     func_path = config["func_path"]
     sys.path.append(func_path)
-    module = __import__("custom_functions")
-    tools = []
-    # 加载JSON文件
-    with open(os.path.join(func_path, "custom_functions.json"), "r") as file:
-        custom_tools = json.load(file)
+    try:
+        module = __import__("custom_functions")
+        tools = []
+        # 加载JSON文件
+        with open(os.path.join(func_path, "custom_functions.json"), "r") as file:
+            custom_tools = json.load(file)
 
-        for c_tool in custom_tools:
-            fn_name = c_tool["function"]["name"]
-            if hasattr(module, fn_name):
-                func = getattr(module, fn_name)
-                fn_schema = create_tool_fn_schema(
-                    fn_name, c_tool["function"]["parameters"]
-                )
-                tool = FunctionTool.from_defaults(
-                    fn=func,
-                    name=fn_name,
-                    fn_schema=fn_schema,
-                    description=c_tool["function"]["description"],
-                )
-                tools.append(tool)
-            else:
-                raise ValueError(
-                    f"Function {fn_name} has not been defined in the custom_functions.py, please define it."
-                )
-    return tools
+            for c_tool in custom_tools:
+                fn_name = c_tool["function"]["name"]
+                if hasattr(module, fn_name):
+                    func = getattr(module, fn_name)
+                    fn_schema = create_tool_fn_schema(
+                        fn_name, c_tool["function"]["parameters"]
+                    )
+                    tool = FunctionTool.from_defaults(
+                        fn=func,
+                        name=fn_name,
+                        fn_schema=fn_schema,
+                        description=c_tool["function"]["description"],
+                    )
+                    tools.append(tool)
+                else:
+                    raise ValueError(
+                        f"Function {fn_name} has not been defined in the custom_functions.py, please define it."
+                    )
+        return tools
+    except Exception:
+        return []
